refactor(storage): split event_log_storage
